requirements.txt
google-genai==0.3.0
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
pyaudio==0.2.14
websockets==13.0
pyautogui>=0.9.54
